Split window state management
